oops

+++ include/gfxalloc.h
@@ -5,6 +5,6 @@
 
 Gfx* Gfx_Open(Gfx* gfx);
 Gfx* Gfx_Close(Gfx* gfx, Gfx* dst);
-void* Gfx_Alloc(Gfx** gfxP, size_t size);
+void* Gfx_Alloc(Gfx** gfxP, u32 size);
 
 #endif
